# frontend/content-hub-frontend/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /search/searchVideo:
    get:
      tags:
      - search-controller
      operationId: searchVideo
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SearchVideoResponseDto"
  /search/searchMovie:
    get:
      tags:
      - search-controller
      operationId: searchMovie
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TmdbSearchMovieDto"
  /search/searchKeyword:
    get:
      tags:
      - search-controller
      operationId: searchKeyword
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
  /search/searchDrama:
    get:
      tags:
      - search-controller
      operationId: searchDrama
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TmdbSearchTvDto"
  /search/searchComics:
    get:
      tags:
      - search-controller
      operationId: searchComics
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: isMainPage
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SearchComicsResponseDto"
  /search/searchAni:
    get:
      tags:
      - search-controller
      operationId: searchAni
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TmdbSearchTvDto"
...
components:
  schemas:
...
    SearchVideoResponseDto:
      type: object
      properties:
        aniResults:
          type: array
          items:
            $ref: "#/components/schemas/TmdbSearchTvResultsDto"
        dramaResults:
          type: array
          items:
            $ref: "#/components/schemas/TmdbSearchTvResultsDto"
        movieResults:
          type: array
          items:
            $ref: "#/components/schemas/TmdbSearchMovieResultsDto"
        page:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
        totalResults:
          type: integer
          format: int32
        aniViewMore:
          type: boolean
        dramaViewMore:
          type: boolean
        movieViewMore:
          type: boolean
    TmdbSearchMovieResultsDto:
      type: object
      properties:
        adult:
          type: boolean
        backdropPath:
          type: string
        genreIds:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: integer
          format: int32
        originalLanguage:
          type: string
        originalName:
          type: string
        overview:
          type: string
        popularity:
          type: number
        posterPath:
          type: string
        voteAverage:
          type: number
        voteCount:
          type: integer
          format: int32
        genreNames:
          type: array
          items:
            type: string
        originalMediaType:
          type: string
        originalTitle:
          type: string
        releaseDate:
          type: string
        title:
          type: string
        video:
          type: boolean
    TmdbSearchTvResultsDto:
      type: object
      properties:
        adult:
          type: boolean
        backdropPath:
          type: string
        genreIds:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: integer
          format: int32
        originalLanguage:
          type: string
        originalName:
          type: string
        overview:
          type: string
        popularity:
          type: number
        posterPath:
          type: string
        voteAverage:
          type: number
        voteCount:
          type: integer
          format: int32
        genreNames:
          type: array
          items:
            type: string
        originalMediaType:
          type: string
        originCountry:
          type: array
          items:
            type: string
        firstAirDate:
          type: string
        name:
          type: string
    TmdbSearchMovieDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
        results:
          type: array
          items:
            $ref: "#/components/schemas/TmdbSearchMovieResultsDto"
        totalPages:
          type: integer
          format: int32
        totalResults:
          type: integer
          format: int32
    TmdbSearchTvDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
        results:
          type: array
          items:
            $ref: "#/components/schemas/TmdbSearchTvResultsDto"
        totalPages:
          type: integer
          format: int32
        totalResults:
          type: integer
          format: int32
    SearchComicsMediaResultDto:
      type: object
      properties:
        adult:
          type: boolean
        backdropPath:
          type: string
        genreIds:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: integer
          format: int32
        originalLanguage:
          type: string
        originalName:
          type: string
        overview:
          type: string
        popularity:
          type: number
        posterPath:
          type: string
        voteAverage:
          type: number
        voteCount:
          type: integer
          format: int32
        genreNames:
          type: array
          items:
            type: string
        originalMediaType:
          type: string
        title:
          type: string
    SearchComicsResponseDto:
      type: object
      properties:
        comicsResults:
          type: array
          items:
            $ref: "#/components/schemas/SearchComicsMediaResultDto"
        page:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
        totalResults:
          type: integer
          format: int32
        comicsViewMore:
          type: boolean
...